BMIを求めよう
