+++ tatebe/java/dojo/java_i/Main.java
@@ -20,25 +20,6 @@ class Main {
     System.out.print("体重：");
     double weight = scanner.nextDouble();
 
-    printData(fullName(firstName, lastName), age, height, weight);
-  }
-
-  public static void printData(String fullName, int age, double height, double weight) {
-    System.out.println("名前は" + fullName + "です");
-  
-    System.out.println("年齢は" + age + "歳です");
-    if (age >= 20) {
-      System.out.println("成年者です");
-    } else {
-      System.out.println("未成年者です");
-    }
-
-    System.out.println("身長は" + height + "mです");
-
-    System.out.println("体重は" + weight + "kgです");
-  }
-
-  public static String fullName(String firstName, String lastName) {
-    return firstName + " " + lastName;
+    Person.printData(Person.fullName(firstName, lastName), age, height, weight, Person.bmi(height, weight));
   }
 }
